refactored details
